+++ router/batchrouter/asyncdestinationmanager/clevertap_segment/utils_test.go
@@ -3,6 +3,8 @@ package clevertapSegment
 import (
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
 )
 
@@ -12,77 +14,94 @@ func TestGetCleverTapEndpoint(t *testing.T) {
 		expectErr bool
 		expectURL string
 	}{
-		{"IN", false, "in1.api.clevertap.com"},
-		{"SINGAPORE", false, "sg1.api.clevertap.com"},
-		{"US", false, "us1.api.clevertap.com"},
-		{"UNKNOWN", true, ""},
+		{region: "IN", expectURL: "in1.api.clevertap.com"},
+		{region: "SINGAPORE", expectURL: "sg1.api.clevertap.com"},
+		{region: "US", expectURL: "us1.api.clevertap.com"},
+		{region: "UNKNOWN", expectErr: true, expectURL: ""},
 	}
 
-	for _, test := range tests {
-		t.Run(test.region, func(t *testing.T) {
-			service := &ClevertapServiceImpl{}
-			endpoint, err := service.getCleverTapEndpoint(test.region)
+	for _, tc := range tests {
+		t.Run(tc.region, func(t *testing.T) {
+			endpoint, err := GetCleverTapEndpoint(tc.region)
 
-			if test.expectErr && err == nil {
-				t.Errorf("expected an error for region %s, got none", test.region)
+			if tc.expectErr {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
 			}
-			if !test.expectErr && err != nil {
-				t.Errorf("did not expect an error for region %s, got: %v", test.region, err)
-			}
-			if endpoint != test.expectURL {
-				t.Errorf("expected URL %s, got %s", test.expectURL, endpoint)
-			}
+			assert.Equal(t, endpoint, tc.expectURL)
 		})
 	}
 }
 
 func TestGetBulkApi(t *testing.T) {
-	service := &ClevertapServiceImpl{}
 	destConfig := DestinationConfig{Region: "IN"}
+	assert.Equal(t, "IN", destConfig.Region)
 
-	bulkApi := service.getBulkApi(destConfig)
-	if bulkApi == nil {
-		t.Fatal("expected a non-nil bulk API service")
-	}
-	if bulkApi.BulkApi == "" {
-		t.Error("expected a non-empty BulkApi URL")
-	}
+	endpoints, err := GetBulkApi(destConfig)
+	assert.Nil(t, err)
+	assert.NotNil(t, endpoints)
+	assert.NotEmpty(t, endpoints.BulkApi)
+	assert.NotEmpty(t, endpoints.NotifyApi)
+	// TODO: Validate endpoints.BulkApi and endpoints.NotifyApi
 }
 
 func TestConvertToConnectionConfig(t *testing.T) {
+	type expected struct {
+		senderName string
+		isErr      bool
+	}
 	tests := []struct {
-		conn      *backendconfig.Connection
-		expectErr bool
+		conn     *backendconfig.Connection
+		expected expected
 	}{
-		{nil, true},
-		{&backendconfig.Connection{
-			SourceID:      "source123",
-			DestinationID: "destination456",
-			Enabled:       true,
-			Config: map[string]interface{}{
-				"Destination": map[string]interface{}{
-					"SchemaVersion": "v1.0",
-					"SegmentName":   "User Segment A",
-					"AdminEmail":    "admin@example.com",
-					"SenderName":    "Rudderstack",
+		{
+			conn: &backendconfig.Connection{
+				SourceID:      "source123",
+				DestinationID: "destination456",
+				Enabled:       true,
+				Config:        nil,
+			},
+			expected: expected{
+				isErr: true,
+			},
+		},
+		{
+			conn: &backendconfig.Connection{
+				SourceID:      "source123",
+				DestinationID: "destination456",
+				Enabled:       true,
+				Config: map[string]interface{}{
+					"Destination": map[string]interface{}{
+						"SchemaVersion": "v1.0",
+						"SegmentName":   "User Segment A",
+						"AdminEmail":    "admin@example.com",
+						"SenderName":    "Rudderstack",
+					},
 				},
 			},
-		}, false},
+			expected: expected{
+				senderName: "Rudderstack",
+				isErr:      false,
+			},
+		},
 	}
 
-	for _, test := range tests {
+	for _, tc := range tests {
 		t.Run("", func(t *testing.T) {
-			service := &ClevertapServiceImpl{}
-			connConfig, err := service.convertToConnectionConfig(test.conn)
+			connConfig, err := ConvertToConnectionConfig(tc.conn)
 
-			if test.expectErr && err == nil {
-				t.Error("expected an error, got none")
+			if tc.expected.isErr {
+				assert.Error(t, err)
+				assert.Equal(t, "", connConfig.Config.Destination.SenderName)
+				return
 			}
-			if !test.expectErr && err != nil {
-				t.Errorf("did not expect an error, got: %v", err)
-			}
-			if !test.expectErr && connConfig.Config.Destination.SenderName != DEFAULT_SENDER_NAME {
-				t.Errorf("expected SenderName to be set to default, got: %s", connConfig.Config.Destination.SenderName)
+			assert.NoError(t, err)
+
+			if connConfig.Config.Destination.SenderName == "" {
+				assert.Equal(t, DEFAULT_SENDER_NAME, connConfig.Config.Destination.SenderName)
+			} else {
+				assert.Equal(t, tc.expected.senderName, connConfig.Config.Destination.SenderName)
 			}
 		})
 	}
